artem polozov

+++ src/main/java/com/example/chocolateshop/repositories/ChocolateRepository.java
@@ -3,5 +3,5 @@ package com.example.chocolateshop.repositories;
 import com.example.chocolateshop.models.Chocolate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChocolateRepository extends JpaRepository<Chocolate, Integer> {
+public interface ChocolateRepository extends JpaRepository<Chocolate, Long> {
 }
